# Julia DataFrame Scan Performance Improvements & TPC-H Tests