--- a/tools/juliapkg/src/data_frame_scan.jl
+++ b/tools/juliapkg/src/data_frame_scan.jl
@@ -2,16 +2,22 @@ using DataFrames
 
 mutable struct DFBindInfo
     df::DataFrame
+    input_columns::Vector
+    scan_types::Vector{Type}
     result_types::Vector{Type}
     scan_functions::Vector{Function}
 
-    function DFBindInfo(df::DataFrame, result_types::Vector{Type}, scan_functions::Vector{Function})
-        return new(df, result_types, scan_functions)
+    function DFBindInfo(df::DataFrame, input_columns::Vector, scan_types::Vector{Type}, result_types::Vector{Type}, scan_functions::Vector{Function})
+        return new(df, input_columns, scan_types, result_types, scan_functions)
     end
 end
 
+function df_input_type(df, entry)
+    return eltype(df[!, entry])
+end
+
 function df_result_type(df, entry)
-    column_type = eltype(df[!, entry])
+    column_type = df_input_type(df, entry)
     if typeof(column_type) == Union
         # remove Missing type from the union
         column_type = Core.Compiler.typesubtract(column_type, Missing, 1)
@@ -47,19 +53,19 @@ function value_to_duckdb(val::T) where {T}
 end
 
 function df_scan_column(
-    df::DataFrame,
+    input_column::Vector{DF_TYPE},
     df_offset::Int64,
     col_idx::Int64,
     result_idx::Int64,
     scan_count::Int64,
     output::DuckDB.DataChunk,
-    ::Type{T}
-) where {T}
-    vector = DuckDB.get_vector(output, result_idx)
-    result_array::Vector{T} = DuckDB.get_array(vector, T)
-    validity = DuckDB.get_validity(vector)
-    input_column = df[!, col_idx]
-    for i in 1:scan_count
+    ::Type{DUCK_TYPE},
+    ::Type{DF_TYPE}
+) where {DUCK_TYPE, DF_TYPE}
+    vector::Vec = DuckDB.get_vector(output, result_idx)
+    result_array::Vector{DUCK_TYPE} = DuckDB.get_array(vector, DUCK_TYPE)
+    validity::ValidityMask = DuckDB.get_validity(vector)
+    for i::Int64 in 1:scan_count
         if input_column[df_offset + i] === missing
             DuckDB.setinvalid(validity, i)
         else
@@ -69,18 +75,18 @@ function df_scan_column(
 end
 
 function df_scan_string_column(
-    df::DataFrame,
+    input_column::Vector{DF_TYPE},
     df_offset::Int64,
     col_idx::Int64,
     result_idx::Int64,
     scan_count::Int64,
     output::DuckDB.DataChunk,
-    ::Type{T}
-) where {T}
-    vector = DuckDB.get_vector(output, result_idx)
-    validity = DuckDB.get_validity(vector)
-    input_column = df[!, col_idx]
-    for i in 1:scan_count
+    ::Type{DUCK_TYPE},
+    ::Type{DF_TYPE}
+) where {DUCK_TYPE, DF_TYPE}
+    vector::Vec = DuckDB.get_vector(output, result_idx)
+    validity::ValidityMask = DuckDB.get_validity(vector)
+    for i::Int64 in 1:scan_count
         if input_column[df_offset + i] === missing
             DuckDB.setinvalid(validity, i)
         else
@@ -106,17 +112,21 @@ function df_bind_function(info::DuckDB.BindInfo)
     df = extra_data[name]
 
     # register the result columns
+    input_columns = Vector()
+    scan_types::Vector{Type} = Vector()
     result_types::Vector{Type} = Vector()
     scan_functions::Vector{Function} = Vector()
     for entry in names(df)
         result_type = df_result_type(df, entry)
         scan_function = df_scan_function(df, entry)
+        push!(input_columns, df[!, entry])
+        push!(scan_types, df_input_type(df, entry))
         push!(result_types, df_julia_type(result_type))
         push!(scan_functions, scan_function)
 
         DuckDB.add_result_column(info, entry, result_type)
     end
-    return DFBindInfo(df, result_types, scan_functions)
+    return DFBindInfo(df, input_columns, scan_types, result_types, scan_functions)
 end
 
 mutable struct DFInitInfo
@@ -137,25 +147,26 @@ function df_scan_function(info::DuckDB.FunctionInfo, output::DuckDB.DataChunk)
     init_info = DuckDB.get_init_info(info, DFInitInfo)
 
     row_count = size(bind_info.df, 1)
-    scan_count = DuckDB.VECTOR_SIZE
+    scan_count::Int64 = DuckDB.VECTOR_SIZE
     if init_info.pos + scan_count >= row_count
         scan_count = row_count - init_info.pos
     end
 
-    result_idx = 1
+    result_idx::Int64 = 1
     for col_idx in init_info.columns
         if col_idx == 0
             result_idx += 1
             continue
         end
         bind_info.scan_functions[col_idx](
-            bind_info.df,
+            bind_info.input_columns[col_idx],
             init_info.pos,
             col_idx,
             result_idx,
             scan_count,
             output,
-            bind_info.result_types[col_idx]
+            bind_info.result_types[col_idx],
+            bind_info.scan_types[col_idx]
         )
         result_idx += 1
     end