--- a/tools/juliapkg/test/runtests.jl
+++ b/tools/juliapkg/test/runtests.jl
@@ -19,7 +19,8 @@ test_files = [
     "test_old_interface.jl",
     "test_all_types.jl",
     "test_decimals.jl",
-    "test_threading.jl"
+    "test_threading.jl",
+    "test_tpch.jl"
 ]
 
 if size(ARGS)[1] > 0