This PR improves the performance of the Julia DataFrame scan by adding a number of extra type annotations to improve type stability.  We also add a test that runs the full TPC-H benchmark using both regular DuckDB versus running the benchmark over Julia DataFrames.

As we can see the type stability improvements drastically improve performance. There is still a small performance penalty over using regular DuckDB tables (and Julia DF scans don't support parallelism yet) - but the performance penalty is not extremely significant (i.e. on the order of 10-30% slower, rather than orders of magnitude).

| Query | Native (1T) | Julia DF (New) | Julia DF (Old) |
|-------|--------|----------------|----------------|
| Q01   | 0.23s  | 0.32s          | 2.35s          |
| Q02   | 0.02s  | 0.06s          | 0.34s          |
| Q03   | 0.09s  | 0.13s          | 1.80s          |
| Q04   | 0.08s  | 0.11s          | 1.34s          |
| Q05   | 0.08s  | 0.15s          | 1.72s          |
| Q06   | 0.04s  | 0.04s          | 1.27s          |
| Q07   | 0.13s  | 0.26s          | 2.14s          |
| Q08   | 0.09s  | 0.15s          | 2.15s          |
| Q09   | 1.22s  | 0.98s          | 3.16s          |
| Q10   | 0.13s  | 0.15s          | 1.62s          |
| Q11   | 0.01s  | 0.01s          | 0.30s          |
| Q12   | 0.16s  | 0.19s          | 1.98s          |
| Q13   | 0.12s  | 0.16s          | 0.46s          |
| Q14   | 0.04s  | 0.05s          | 1.45s          |
| Q15   | 0.06s  | 0.09s          | 2.74s          |
| Q16   | 0.03s  | 0.04s          | 0.15s          |
| Q17   | 0.10s  | 0.14s          | 1.78s          |
| Q18   | 0.24s  | 0.60s          | 2.23s          |
| Q19   | 0.19s  | 0.27s          | 1.91s          |
| Q20   | 0.09s  | 0.11s          | 1.52s          |
| Q21   | 0.32s  | 0.44s          | 4.00s          |
| Q22   | 0.02s  | 0.03s          | 0.15s          |